cleanup, almost complete with new architecture refactor
