+++ app/src/main/java/csc_cccix/geocracy/game/IStateMachine.java
@@ -1,6 +1,6 @@
 package csc_cccix.geocracy.game;
 
-import csc_cccix.geocracy.old_states.GameEvent;
+import csc_cccix.geocracy.game.ui_states.GameEvent;
 
 public abstract class IStateMachine {
