+++ app/src/main/java/csc_cccix/geocracy/game/ui_states/GameInfoVisibleState.java
@@ -5,7 +5,6 @@ import android.util.Log;
 import csc_cccix.geocracy.fragments.GameInfoFragment;
 import csc_cccix.geocracy.game.IState;
 import csc_cccix.geocracy.game.IStateMachine;
-import csc_cccix.geocracy.old_states.GameEvent;
 
 public class GameInfoVisibleState extends IGameplayState {
