+++ app/src/main/java/csc_cccix/geocracy/game/ui_states/IGameplayState.java
@@ -4,7 +4,6 @@ import android.util.Log;
 
 import csc_cccix.geocracy.game.IState;
 import csc_cccix.geocracy.game.IStateMachine;
-import csc_cccix.geocracy.old_states.GameEvent;
 
 public abstract class IGameplayState extends IState {
