+++ app/src/main/java/csc_cccix/geocracy/game/ui_states/IntentToAttackState.java
@@ -3,7 +3,6 @@ package csc_cccix.geocracy.game.ui_states;
 import android.util.Log;
 
 import csc_cccix.geocracy.game.IStateMachine;
-import csc_cccix.geocracy.old_states.GameEvent;
 import csc_cccix.geocracy.world.Territory;
 
 public class IntentToAttackState extends IGameplayState {
