+++ app/src/main/java/csc_cccix/geocracy/game/ui_states/PlaceReinforcementsState.java
@@ -3,15 +3,12 @@ package csc_cccix.geocracy.game.ui_states;
 import android.util.Log;
 import android.widget.Toast;
 
-import java.util.HashSet;
-
 import csc_cccix.geocracy.Util;
 import csc_cccix.geocracy.fragments.DistributeTroopsDetailFragment;
 import csc_cccix.geocracy.game.Game;
 import csc_cccix.geocracy.game.HumanPlayer;
 import csc_cccix.geocracy.game.IStateMachine;
 import csc_cccix.geocracy.game.Player;
-import csc_cccix.geocracy.old_states.GameEvent;
 import csc_cccix.geocracy.world.Territory;
 import es.dmoral.toasty.Toasty;
