--- app/src/main/java/csc_cccix/geocracy/old_states/BattleResultDetails.java
@@ -1,14 +1,0 @@
-package csc_cccix.geocracy.old_states;
-
-class BattleResultDetails {
-
-    public int attackerArmiesLost;
-    public int defenderArmiesLost;
-
-
-    public BattleResultDetails(int attackerArmiesLost, int defenderArmiesLost) {
-        this.attackerArmiesLost = attackerArmiesLost;
-        this.defenderArmiesLost = defenderArmiesLost;
-    }
-
-}
